Move things around one class per file `PublishEnvelope` lives in `JustSaying.Messaging` with the `IMessagePublisher` where it is used. get rid of all the `using JustSaying.Models;`
