+++ JustSaying.IntegrationTests/FluentNotificationStackTestBase.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Reflection;
 using System.Threading.Tasks;
 using Amazon;
 using Amazon.SimpleNotificationService.Model;
