+++ JustSaying.UnitTests/JustSayingBus/WhenPublishingMessages.cs
@@ -1,7 +1,6 @@
 using System.Threading;
 using System.Threading.Tasks;
 using JustSaying.Messaging;
-using JustSaying.Models;
 using JustSaying.TestingFramework;
 using NSubstitute;
 using Xunit;
