+++ JustSaying.UnitTests/JustSayingBus/WhenRegisteringMessageHandlers.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Threading;
 using System.Threading.Tasks;
 using JustSaying.Messaging;
 using JustSaying.Messaging.MessageHandling;
