+++ JustSaying.UnitTests/JustSayingBus/WhenRegisteringPublishers.cs
@@ -2,7 +2,6 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using JustSaying.Messaging;
-using JustSaying.Models;
 using JustSaying.TestingFramework;
 using NSubstitute;
 using Shouldly;
