+++ JustSaying.UnitTests/JustSayingFluently/RegisteringPublishers/WhenAddingPublishers.cs
@@ -1,4 +1,3 @@
-using System.Threading.Tasks;
 using JustBehave;
 using JustSaying.AwsTools;
 using JustSaying.TestingFramework;
