+++ JustSaying.UnitTests/JustSayingFluentlyTestBase.cs
@@ -4,7 +4,6 @@ using JustSaying.AwsTools.QueueCreation;
 using JustBehave;
 using Microsoft.Extensions.Logging;
 using NSubstitute;
-using Xunit;
 
 namespace JustSaying.UnitTests
 {
