+++ JustSaying/AwsTools/AsyncLazy.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Runtime.CompilerServices;
 using System.Threading.Tasks;
 
 namespace JustSaying.AwsTools
