+++ JustSaying/AwsTools/MessageHandling/LocalTopicArnProvider.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Threading.Tasks;
 using Amazon.SimpleNotificationService;
