+++ JustSaying/AwsTools/MessageHandling/SqsNotificationListener.cs
@@ -1,8 +1,6 @@
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
-using System.Globalization;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using Amazon.SQS;
