+++ JustSaying/AwsTools/MessageHandling/SqsQueueBase.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using System.Globalization;
 using System.Net;
 using System.Threading.Tasks;
 using Amazon;
