+++ JustSaying/AwsTools/MessageHandling/SqsQueueByName.cs
@@ -1,5 +1,4 @@
 using System.Collections.Generic;
-using System.Globalization;
 using System.Net;
 using System.Threading.Tasks;
 using Amazon;
