+++ JustSaying/Messaging/MessageHandling/IMessageLockAsync.cs
@@ -1,6 +1,4 @@
 using System;
-using System.Collections.Generic;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace JustSaying.Messaging.MessageHandling
